[FIX] #160 fix magerun2 stop-file detection
the magerun2 stop-file was not detected because the wrong file-name was
looked for: instead of looking for ".n98-magerun2" it was looked for
"n98-magerun2.yaml".

the fix is to look for the correct filename ".n98-magerun2".

the stop-file-name is now the dot (".") followed by the basename of the
_customConfigFilename with the last extension parameter removed
(PATHINFO_FILENAME).

additionally:

- bump version 1.0.1